Working on text flow indexing.

+++ src/Rendering/Assets/Text/Text.cpp
@@ -145,7 +145,7 @@ namespace eyegui
         positions.reserve(mVertexCount);
         std::vector<glm::vec2> textureCoordinates;
         textureCoordinates.reserve(mVertexCount);
-        for(int i = 0; i < mVertexCount; i++)
+        for(int i = 0; i < (int)mVertexCount; i++)
         {
             positions.push_back(vertices.at(i).first);
             textureCoordinates.push_back(vertices.at(i).second);
@@ -168,7 +168,7 @@ namespace eyegui
         RenderWord word;
         word.spVertices = std::shared_ptr<RenderWordVertices>(new RenderWordVertices);
         word.lettersXOffsets.reserve(rContent.size() + 1); // first position is zero
-        word.lettersXOffsets.push_back(0.f);
+        word.lettersXOffsets.push_back(0);
 
         // Fill word with data
         float xPixelPen = 0;
@@ -201,7 +201,7 @@ namespace eyegui
             xPixelPen += scale * (float)pGlyph->advance.x;
 
 			// Push back x offset of letters
-			word.lettersXOffsets.push_back(xPixelPen);
+			word.lettersXOffsets.push_back((int)xPixelPen);
 
             // Fill into data blocks
             word.spVertices->push_back(std::make_pair(positionA, textureCoordinateA));
